requirements.txt
# Core LLM Frameworks
llama-index-core==0.13.5
llama-index-llms-llama-cpp==0.5.1
llama-index-embeddings-huggingface==0.6.0
llama-index-vector-stores-chroma==0.5.2
langchain==1.2.0
langchain-community==0.4.1

# Inference Backends

llama-cpp-python==0.3.16
#vLLM will be installed separately on PC

# Vector Store
# chromadb==1.3.7

# Embeddings
sentence-transformers==3.4.1
transformers==4.57.3
tokenizers==0.22.1 

# Observability
arize-phoenix==12.25.1
openinference-instrumentation-langchain==0.1.56
openinference-instrumentation-llama-index==4.3.9


# Document Processing
PyMuPDF==1.25.2      

# UI
streamlit==1.41.1
plotly==5.24.1

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
#pydantic==2.12.2
#pydantic-settings==2.12.0
#numpy==1.26.4 -> already have by pytorch
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0


# Install outside requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu